[Renderer] Allow indepedent toc access.
The previous solution just stored a toc update request in the list
of operations to be done after rendering. As the toc is one of the
first elements in a new document the update request is also one of
first operations inside this list. This is okay as long as none of
the following operations triggers a change for the headings page
numbering.

The new version always triggers a toc update after everything else
is done. If required it should also be possible to split the list
of operations stored in `afterRendering` into `before toc update`
and `after toc update`.