[Renderer] Avoid space after inline source on paragraph end.
The old rendering hook adds a new space at the beginning of a new
paragraph if the previous one ended with some inline source code.

Reason was a missing check if the next word is at the beginning
of a new paragraph inside the custom `skip_if` lambda inside the
method which handles inline source insertion.