[Parser] Add new parsing function for table cells.
The new `parseInlineDocument` works almost like the existing function
`parseDocument` but allows to pass an initial source position.

If an error appears this initial source position is used to create
the actual error message which should result in way more specific
error messages as before for all kind of child parsers which are
used to parse parts of a bigger file (e.g. table cells).