[Renderer] Remove redundant space in inline quotes.
Delete the internal state storing the last known item during the
initialization of an inline quote to avoid a redundant space if
the first element inside the inline quotation container is not
a simple word but some kind of content creation tag.

This is necessary as the inline quote works like a container and
therefore the renderer still references the last element in front
of the inline quote instead of the container itself.

If the first child element relies on `needSpace()` to determine
if an additional space should be injected it would still base this
decision on the previous element, even if the handler itself already
injected a quotation mark into the text.