[Renderer] Extract alignment and v_offset handling from _insertImage().
Since images are always injected into a frame to group them with their
related caption, the alignment handling shouldn't be part of the image
but of the surrounding frame element. Especially as inline images never
utilized the alignment parameter. The same is true for the v_offset,
which is only used by inline images.

Therefore both parameter got removed from _insertImage() and the effected
parts of the code updated.

Attention: This required to change the value used to change the frames
           horizontal orientation to left. Original: 1, Now: 3