[Renderer] Allow to restore character properties.
As some of the character properties like 'CharStyleName' potentially
influence any of the other properties it isn't enough to just restore
it to its original value (e.g. the default value of 'CharStyleName'
clears the 'CharWeight' back to 100.0).

Therefore the new approach allows to restore not only the current
property but also any other one, which is writable and not empty.

For some currently unknown reasons this isn't possible for each property
it was therefore possible to introduce some special case handling into
the restoration function.