[Parser] Abort if input contains non-printable characters.
Denies to parse any input, which includes non-printable characters,
as the underlying LibreOffice renderer wouldn't be able to include
those properly into the final document.

It seems like this changeset aborts properly if a such a chracter got
encountered, but I was unable to come up with a solution, that would
actually allow to print a human understandable error message.