[Renderer] Don't restore unchanged properties.
Should fix a regression which prevents proper heading formatting,
e.g. after inline source elements.

Seems like this regression is related to one of the properties we
restore to default but didn't change in the first place. Therefore,
Libre Office starts to assume that there is already a custom text
style applied which sometimes prevent the usage of other Fonts.