Add support for source code lines
We add three more properties to the `source element` of the format

    [source firstnumber=N, range="a-b,c-d", numbers=Bool]

The `firstnumber` determines how we remap the first actual input line.
By default this is the identity, that is, the first input line is set as
first line in the output: `firstnumber=0`.

The range specifies what parts of the code we want to print.
It depends on the value of `firstnumber`, that is, if we use `firstnumber=10` and
`range 11-14` it will print the 2st to 6nd (thx @aluc) line of the input.
By default all lines are printed.

Finally, we can turn line numbers on or off.